feat(qr-generator.jsx): add QR code generator component
This commit adds a new file `qr-generator.jsx` which contains a React component for generating QR codes. The component imports the necessary dependencies, including `react-qr-code` and a custom button component.

The `QrGenerator` component uses the `useState` hook to manage a `tableCounter` state variable. It renders a QR code using the `react-qr-code` component, with a size of 150 pixels and a value of `'test' + tableCounter`. The QR code is displayed within a `<div>` element and is styled to have a maximum width of 100% and auto height.

Additionally, the component renders a button with the text "Done" and an `onClick` event handler that increments the `tableCounter` state variable by 1 when clicked.

This component can be used to generate QR codes with dynamic values and provides a convenient way to increment the `tableCounter` value.