feat(package-lock.json): add react-qr-code package to project dependencies
The react-qr-code package was added to the project dependencies to enable the generation of QR codes in the application. This package provides a React component for rendering QR codes. The version added is 2.0.12.

Additionally, the qr.js package was added as a dependency of react-qr-code. This package is required by react-qr-code for generating QR codes.

The react-qr-code package has peer dependencies on react and react-native-svg. The version range for react is "^16.x || ^17.x || ^18.x", and the version range for react-native-svg is "*". These peer dependencies are required for the proper functioning of react-qr-code.

The react-native-svg peer dependency is marked as optional in the peerDependenciesMeta section, indicating that it is not strictly required but recommended for better compatibility with react-qr-code.